Use most recent plugin bom release
The release reports generated by sites like https://mvrepository.com
seem to apply the same version number parsing rules as Apache Maven.
That causes those sites to see version numbers that contain the letter
'a' or the letter 'b' followed by an alphanumeric character as a
designator for an alpha or a beta release.

See https://github.com/jenkinsci/incrementals-tools/pull/24 for the
rationale for that specific change and
https://github.com/jenkinsci/bom#readme for the details of the bom and
https://github.com/jenkinsci/jep/blob/master/jep/229/README.adoc for
the version number details.